# backend/turismo-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Turismo Necochea",
        "description": "Documentación de la API para la aplicación de Turismo de Necochea",
        "contact": {
            "email": "tudai@alumnos.exa.unicen.edu.ar"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/alojamientos": {
            "get": {
                "tags": [
                    "Alojamientos"
                ],
                "summary": "Obtener todos los alojamientos",
                "operationId": "5a3e7174cb15d6c117ba171398cd62e1",
                "responses": {
                    "200": {
                        "description": "Lista de alojamientos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Alojamiento"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Alojamientos"
                ],
                "summary": "Crear un nuevo alojamiento",
                "operationId": "dff57dafa4fe7cbf4bcf339386bbdd5c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlojamientoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alojamiento creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alojamiento"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/alojamientos/{id}": {
            "get": {
                "tags": [
                    "Alojamientos"
                ],
                "summary": "Obtener un alojamiento por ID",
                "operationId": "955635c26c9b11f48bda4fe985036085",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del alojamiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del alojamiento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alojamiento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alojamiento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Alojamientos"
                ],
                "summary": "Actualizar un alojamiento por ID",
                "operationId": "95f1f2196ca5cec2550cd96a3154ecf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del alojamiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlojamientoUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alojamiento actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Alojamiento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alojamiento no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alojamientos"
                ],
                "summary": "Eliminar un alojamiento por ID",
                "operationId": "18e845b88a2155037c890f09bfc141d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del alojamiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alojamiento eliminado correctamente"
                    },
                    "404": {
                        "description": "Alojamiento no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Alojamiento": {
                "title": "Alojamiento",
                "description": "Modelo de un alojamiento",
                "properties": {
                    "id": {
                        "description": "ID único del alojamiento",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "description": "Nombre del alojamiento",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del alojamiento",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Número de teléfono",
                        "type": "string",
                        "nullable": true
                    },
                    "redesSociales": {
                        "description": "URL de redes sociales",
                        "type": "string",
                        "nullable": true
                    },
                    "paginaWeb": {
                        "description": "Página web del alojamiento",
                        "type": "string",
                        "nullable": true
                    },
                    "mail": {
                        "description": "Correo electrónico de contacto",
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "mascotas": {
                        "description": "Indica si se permiten mascotas",
                        "type": "boolean"
                    },
                    "periodoApertura": {
                        "description": "Período de apertura (ej. 'Todo el año', 'Temporada alta')",
                        "type": "string",
                        "nullable": true
                    },
                    "tipo": {
                        "description": "Tipo de alojamiento (ej. 'Hotel', 'Cabaña')",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "AlojamientoRequest": {
                "title": "AlojamientoRequest",
                "description": "Esquema para la creación de un alojamiento",
                "required": [
                    "nombre",
                    "direccion",
                    "mascotas",
                    "tipo"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Hotel Central"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Calle 10 N° 500"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2262554433",
                        "nullable": true
                    },
                    "redesSociales": {
                        "type": "string",
                        "example": "http://instagram.com/hotelcentral",
                        "nullable": true
                    },
                    "paginaWeb": {
                        "type": "string",
                        "example": "http://hotelcentral.com",
                        "nullable": true
                    },
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "reservas@hotelcentral.com",
                        "nullable": true
                    },
                    "mascotas": {
                        "type": "boolean",
                        "example": false
                    },
                    "periodoApertura": {
                        "type": "string",
                        "example": "Todo el año",
                        "nullable": true
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Hotel"
                    }
                },
                "type": "object"
            },
            "AlojamientoUpdateRequest": {
                "title": "AlojamientoUpdateRequest",
                "description": "Esquema para la actualización de un alojamiento",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Hotel del Parque",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. 79 N° 1200",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "2262778899",
                        "nullable": true
                    },
                    "redesSociales": {
                        "type": "string",
                        "example": "http://facebook.com/hoteldelparque",
                        "nullable": true
                    },
                    "paginaWeb": {
                        "type": "string",
                        "example": "http://hoteldelparque.com",
                        "nullable": true
                    },
                    "mail": {
                        "type": "string",
                        "format": "email",
                        "example": "info@hoteldelparque.com",
                        "nullable": true
                    },
                    "mascotas": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    },
                    "periodoApertura": {
                        "type": "string",
                        "example": "Temporada alta",
                        "nullable": true
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Cabaña",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}